refactor(sandbox): drive sandbox context note from unified resolver
