+++ test/hooks/sandbox-message.test.ts
@@ -1,20 +1,20 @@
 import { describe, test, expect, vi } from 'vitest'
 import { createSandboxMessageHook } from '../../src/hooks/sandbox-message'
-import { SANDBOX_CONTEXT_NOTE } from '../../src/loop/prompts'
+import { SANDBOX_CONTEXT_NOTE } from '../../src/sandbox/context'
+import type { SandboxContext } from '../../src/sandbox/context'
 import type { Logger } from '../../src/types'
 
 const logger = { log: vi.fn(), error: vi.fn(), debug: vi.fn() } as unknown as Logger
 
-function makeHook(resolved: unknown) {
-  const sessionLoopResolver = {
-    resolveActiveLoopForSession: vi.fn(async () => resolved),
-  } as any
-  return createSandboxMessageHook({ sessionLoopResolver, logger })
+const context = { containerName: 'forge-x', hostDir: '/w', mounts: [] } as unknown as SandboxContext
+
+function makeHook(resolved: SandboxContext | null) {
+  return createSandboxMessageHook({ resolveSandboxForSession: async () => resolved, logger })
 }
 
 describe('createSandboxMessageHook (chat.system.transform)', () => {
-  test('appends sandbox context note to system prompt for an active sandbox loop session', async () => {
-    const hook = makeHook({ loopName: 'l', active: true, sandbox: true })
+  test('appends sandbox context note when the session resolves to a container', async () => {
+    const hook = makeHook(context)
     const output = { system: ['base system prompt'] }
 
     await hook({ sessionID: 'ses_1' }, output)
@@ -23,25 +23,7 @@ describe('createSandboxMessageHook (chat.system.transform)', () => {
     expect(output.system).toHaveLength(2)
   })
 
-  test('does not append for a non-sandbox (worktree-only) loop session', async () => {
-    const hook = makeHook({ loopName: 'l', active: true, sandbox: false })
-    const output = { system: ['base'] }
-
-    await hook({ sessionID: 'ses_1' }, output)
-
-    expect(output.system).toEqual(['base'])
-  })
-
-  test('does not append for an inactive loop session', async () => {
-    const hook = makeHook({ loopName: 'l', active: false, sandbox: true })
-    const output = { system: ['base'] }
-
-    await hook({ sessionID: 'ses_1' }, output)
-
-    expect(output.system).toEqual(['base'])
-  })
-
-  test('does not append for a session that resolves to no loop', async () => {
+  test('does not append when the session resolves to no sandbox', async () => {
     const hook = makeHook(null)
     const output = { system: ['base'] }
 
@@ -51,24 +33,21 @@ describe('createSandboxMessageHook (chat.system.transform)', () => {
   })
 
   test('no-ops when sessionID is missing', async () => {
-    const resolver = vi.fn(async () => ({ active: true, sandbox: true }))
-    const hook = createSandboxMessageHook({
-      sessionLoopResolver: { resolveActiveLoopForSession: resolver } as any,
-      logger,
-    })
+    const resolveSandboxForSession = vi.fn(async () => context)
+    const hook = createSandboxMessageHook({ resolveSandboxForSession, logger })
     const output = { system: ['base'] }
 
     await hook({}, output)
 
-    expect(resolver).not.toHaveBeenCalled()
+    expect(resolveSandboxForSession).not.toHaveBeenCalled()
     expect(output.system).toEqual(['base'])
   })
 
   test('swallows resolver errors without throwing or appending', async () => {
-    const sessionLoopResolver = {
-      resolveActiveLoopForSession: vi.fn(async () => { throw new Error('boom') }),
-    } as any
-    const hook = createSandboxMessageHook({ sessionLoopResolver, logger })
+    const hook = createSandboxMessageHook({
+      resolveSandboxForSession: async () => { throw new Error('boom') },
+      logger,
+    })
     const output = { system: ['base'] }
 
     await expect(hook({ sessionID: 'ses_1' }, output)).resolves.toBeUndefined()
